% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{demo_scale}
\alias{demo_scale}
\title{Demonstrate psychrometric label and scale functions}
\usage{
demo_scale(x, ...)
}
\arguments{
\item{x}{A vector of data}

\item{...}{Other arguments pass to scale functions}
}
\value{
A ggplot object demonstrating the supplied scale settings.
}
\description{
This helper builds a compact ggplot2 scale preview for label and scale
functions.
}
\examples{
demo_scale(0:10, labels = scales::label_number())

}
