% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.R, R/coord-psychro.R
\docType{data}
\name{ggpsychro-ggproto}
\alias{ggpsychro-ggproto}
\alias{CoordPsychro}
\title{Base ggproto classes for ggpsychro}
\description{
Base ggproto classes for ggpsychro
}
\seealso{
ggplot2::ggproto
}
\keyword{datasets}
\keyword{internal}
