% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro-package.R
\docType{package}
\name{ggpsychro-package}
\alias{ggpsychro-package}
\title{ggpsychro: psychrometric charts with ggplot2}
\description{
\code{ggpsychro} extends ggplot2 with psychrometric coordinates, grids, data
layers, zones, process lines, and thermal-comfort overlays.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hongyuanjia.github.io/ggpsychro/}
  \item \url{https://github.com/hongyuanjia/ggpsychro}
  \item Report bugs at \url{https://github.com/hongyuanjia/ggpsychro/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID}) [copyright holder]

}
