% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro.R
\name{is.ggpsychro}
\alias{is.ggpsychro}
\title{Reports whether x is a ggplot object}
\usage{
is.ggpsychro(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A single logical value.
}
\description{
Reports whether x is a ggplot object
}
\examples{
is.ggpsychro(ggpsychro())
is.ggpsychro(ggplot2::ggplot())

}
\keyword{internal}
