% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro.R
\name{is_ggpsychro}
\alias{is_ggpsychro}
\title{Test for ggpsychro plots}
\usage{
is_ggpsychro(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A single logical value.
}
\description{
Test for ggpsychro plots
}
\examples{
is_ggpsychro(ggpsychro())
is_ggpsychro(ggplot2::ggplot())

}
