% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_psychro}
\alias{theme_psychro}
\alias{theme_grey_psychro}
\alias{theme_gray_psychro}
\alias{theme_psychro_ashrae}
\alias{theme_psychro_minimal}
\title{Custom theme for psychrometric chart.}
\usage{
theme_grey_psychro(
  base_size = 11,
  base_family = "",
  header_family = NULL,
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ink = "black",
  paper = "white",
  accent = "#3366FF"
)

theme_gray_psychro(
  base_size = 11,
  base_family = "",
  header_family = NULL,
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ink = "black",
  paper = "white",
  accent = "#3366FF"
)

theme_psychro(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_psychro_ashrae(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_psychro_minimal(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{header_family}{font family for titles and headers. The default, \code{NULL},
uses theme inheritance to set the font. This setting affects axis titles,
legend titles, the plot title and tag text.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{ink, paper, accent}{colour for foreground, background, and accented elements respectively.}
}
\value{
A ggplot2 theme.

\code{theme_psychro_ashrae()} and \code{theme_psychro_minimal()} are inspired by
psychrochart's ASHRAE and minimal chart styles.
}
\description{
Custom theme for psychrometric chart.
}
\examples{
ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 30)) +
    theme_grey_psychro()

ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 30)) +
    theme_gray_psychro()

ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 30)) +
    theme_psychro()

ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 30)) +
    theme_psychro_ashrae()

ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 30)) +
    theme_psychro_minimal()

}
\seealso{
\itemize{
\item \href{https://github.com/azogue/psychrochart/blob/master/psychrochart/chart_styles/ashrae_chart_style.json}{psychrochart ASHRAE style}
\item \href{https://github.com/azogue/psychrochart/blob/master/psychrochart/chart_styles/minimal_chart_style.json}{psychrochart minimal style}
}
}
\author{
Hongyuan Jia
}
