% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{drybulb_trans}
\alias{drybulb_trans}
\alias{humratio_trans}
\alias{relhum_trans}
\alias{wetbulb_trans}
\alias{vappres_trans}
\alias{specvol_trans}
\alias{enthalpy_trans}
\title{Create transformation objects for psychrometric chart}
\usage{
drybulb_trans(units = "SI")

humratio_trans(units = "SI")

relhum_trans(units = "SI")

wetbulb_trans(units = "SI")

vappres_trans(units = "SI")

specvol_trans(units = "SI")

enthalpy_trans(units = "SI")
}
\arguments{
\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}.}
}
\description{
Create transformation objects for psychrometric chart
}
\examples{
plot(drybulb_trans("SI"), xlim = c(0, 5))
plot(humratio_trans("SI"), xlim = c(0, 1000))
plot(relhum_trans("SI"), xlim = c(0, 1))
plot(wetbulb_trans("SI"), xlim = c(-50, 40))
plot(vappres_trans("SI"), xlim = c(1000, 4000))
plot(specvol_trans("SI"), xlim = c(0.8, 1))
}
