% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-psychro.R
\name{coord_psychro}
\alias{coord_psychro}
\title{Psychrometric coordinates}
\usage{
coord_psychro(
  tdb_lim = NULL,
  hum_lim = NULL,
  altitude = NULL,
  units = NULL,
  mollier = NULL,
  expand = FALSE,
  default = TRUE,
  clip = "on"
)
}
\arguments{
\item{tdb_lim}{A numeric vector of length-2 indicating the dry-bulb
temperature limits. Should be in range
\verb{[-50, 100]} degree_C [SI] or
\verb{[-58, 212]} degree_F [IP].
If \code{NULL}, no grid lines will be plotted when no data and aesthetics
are added, just like normal ggplot behaviour. Default: \code{NULL}.}

\item{hum_lim}{A numeric vector of length-2 indicating the humidity ratio
limits. Should be in range
\verb{[0, 60]} g_H20 kg_Air-1 [SI] or
\verb{[0, 350]} gr_H20 lb_Air-1 [IP].
If \code{NULL}, no grid lines will be plotted when no data and aesthetics
are added, just like normal ggplot behaviour. Default: \code{NULL}.}

\item{altitude}{A single number of altitude in m [SI] or ft [IP]. Default:
\code{0}.}

\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}.}

\item{mollier}{If \code{TRUE}, a Mollier chart will be created instead of a
psychrometric chart. Default: \code{FALSE}.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{default}{Is this the default coordinate system? If \code{FALSE} (the default),
then replacing this coordinate system with another one creates a message alerting
the user that the coordinate system is being replaced. If \code{TRUE}, that warning
is suppressed.}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. In most cases, the default of \code{"on"} should not be changed,
as setting \code{clip = "off"} can cause unexpected results. It allows
drawing of data points anywhere on the plot, including in the plot margins. If
limits are set via \code{xlim} and \code{ylim} and some data points fall outside those
limits, then those data points may show up in places such as the axes, the
legend, the plot title, or the plot margins.}
}
\description{
Psychrometric coordinates
}
