% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{demo_scale}
\alias{demo_scale}
\title{Demonstrate scales functions with ggplot2 code}
\usage{
demo_scale(x, ...)
}
\arguments{
\item{x}{A vector of data}

\item{...}{Other arguments pass to scale functions}
}
\description{
This function generates ggplot2 code needed to use scales functions for real
code.
}
