% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.R
\name{element_polygon}
\alias{element_polygon}
\title{Polygon theme element for psychrometric chart panels}
\usage{
element_polygon(
  fill = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{fill}{Fill colour.}

\item{colour, color}{Border colour. \code{color} is an alias for \code{colour}.}

\item{size}{Border width.}

\item{linetype}{Border linetype.}

\item{inherit.blank}{Whether this element inherits from blank elements.}
}
\value{
A ggplot2 theme element.
}
\description{
\code{element_polygon()} is used by ggpsychro-specific theme elements such as
\code{psychro.panel.background} and \code{psychro.panel.mask}.
}
\examples{
ggpsychro(tdb_lim = c(0, 40), hum_lim = c(0, 25)) +
    ggplot2::theme(
        psychro.panel.background = element_polygon(
            fill = "#F7F7F2",
            colour = "grey60",
            size = 0.4
        )
    )

}
