% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-grid.R
\name{geom_grid_relhum}
\alias{geom_grid_relhum}
\alias{geom_grid_wetbulb}
\alias{geom_grid_vappres}
\alias{geom_grid_specvol}
\alias{geom_grid_enthalpy}
\title{Draw grid lines of constant psychrometric properties}
\usage{
geom_grid_relhum(
  mapping = NULL,
  data = NULL,
  n = 201,
  label_loc = 0.95,
  label_parse = FALSE,
  units = waiver(),
  pres = waiver(),
  ...,
  na.rm = FALSE
)

geom_grid_wetbulb(
  mapping = NULL,
  data = NULL,
  label_loc = 0.1,
  label_parse = TRUE,
  units = waiver(),
  pres = waiver(),
  ...,
  na.rm = FALSE
)

geom_grid_vappres(
  mapping = NULL,
  data = NULL,
  label_loc = 0.5,
  label_parse = FALSE,
  units = waiver(),
  pres = waiver(),
  ...,
  na.rm = FALSE
)

geom_grid_specvol(
  mapping = NULL,
  data = NULL,
  label_loc = 0.95,
  label_parse = TRUE,
  units = waiver(),
  pres = waiver(),
  ...,
  na.rm = FALSE
)

geom_grid_enthalpy(
  mapping = NULL,
  data = NULL,
  label_loc = 0.95,
  label_parse = TRUE,
  units = waiver(),
  pres = waiver(),
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{n}{Number of points to interpolate along. Only used in
\code{geom_grid_relhum()}.}

\item{label_loc}{A single number in range (0, 1) indicating label position
relative to the line. The default values aim to reduce overlappings
among differnent psychrometric property lines, but you can change them
if needed:
\itemize{
\item \code{0.95} for constant relative humidity lines
\item \code{0.10} for constant wet-bulb temperature lines
\item \code{0.50} for constant water vapor pressure lines
\item \code{0.95} for constant specific volume lines
\item \code{0.95} for constant enthalpy lines
}}

\item{label_parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{units}{A single string indicating the units sytem to use. Should be
either \code{"SI"} or \verb{"IP" or }waiver()\verb{which uses the value from the parent plot. Default:}waiver()`}

\item{pres}{A single number indicating the atmosphere pressure in Pa [SI] or
Psi [IP]. If \code{waiver()}, the pressure calculated from the parent
plot's altitude value will be used. Default: \code{waiver()}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
\verb{geom_grid_*()} geoms draw grid line of constant psychrometric properties,
including relative humidity, wet-bulb temperature, water vapor pressure,
specific volume and enthalpy, based on current psychrometric chart's dry-bulb
temperature range and humidity ratio range.
}
\details{
\itemize{
\item \code{geom_grid_relhum()} for relative humidity grid in range [0, 100] in \%
\item \code{geom_grid_wetbulb()} for wet-bulb temperature grid in °F [IP] or °C [SI]
\item \code{geom_grid_vappres()} for partial pressure grid of water vapor in Psi [IP]
or Pa [SI]
\item \code{geom_grid_specvol()} for specific volume grid in ft3 lb-1 of dry air [IP] or
in m3 kg-1 of dry air [SI]
\item \code{geom_grid_enthalpy()} for moist air enthalpy grid in Btu lb-1 [IP] or kJ kg-1
}

Each \verb{geom_grid_*()} comes along with a corresponding
\code{\link[=scale_relhum]{scale_*()}} function for customizing scale
properties, including breaks, labels and etc.

For each psychrometric properties, the maximum and minimum value is
calculated based on the ranges of dry-bulb temperature and humidity ratio in
the coordinate system.
}
\section{Alignment}{


You can modify text alignment with the \code{vjust} and \code{hjust}
aesthetics. These can either be a number between 0 (right/bottom) and
1 (top/left) or a character (\code{"left"}, \code{"middle"}, \code{"right"}, \code{"bottom"},
\code{"center"}, \code{"top"}). There are two special alignments: \code{"inward"} and
\code{"outward"}. Inward always aligns text towards the center, and outward
aligns it away from the center.
}

\section{Aesthetics}{


\verb{geom_grid_*()} understands the following aesthetics.
\itemize{
\item \code{color}
\item \code{size}
\item \code{linetype}
\item \code{alpha}
\item \code{label.colour}
\item \code{label.size}
\item \code{label.angle}
\item \code{label.hjust}
\item \code{label.vjust}
\item \code{label.alpha}
\item \code{label.family}
\item \code{label.fontface}
\item \code{label.lineheight}
}
}

