% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-grid.R
\name{geom_line_sat}
\alias{geom_line_sat}
\alias{geom_grid_relhum}
\alias{geom_grid_wetbulb}
\alias{geom_grid_vappres}
\alias{geom_grid_specvol}
\alias{geom_grid_enthalpy}
\title{Draw constant psychrometric properties grids}
\usage{
geom_line_sat(
  mapping = NULL,
  data = NULL,
  n = 201,
  ...,
  na.rm = FALSE,
  inherit.aes = TRUE
)

geom_grid_relhum(
  mapping = NULL,
  data = NULL,
  step = 10,
  n = 201,
  label = NULL,
  label_loc = NULL,
  label_fun = NULL,
  label_parse = FALSE,
  label_style = NULL,
  check_overlap = FALSE,
  ...,
  na.rm = FALSE,
  inherit.aes = TRUE
)

geom_grid_wetbulb(
  mapping = NULL,
  data = NULL,
  step = 10,
  label = NULL,
  label_loc = NULL,
  label_fun = NULL,
  label_parse = FALSE,
  label_style = NULL,
  check_overlap = FALSE,
  ...,
  na.rm = FALSE,
  inherit.aes = TRUE
)

geom_grid_vappres(
  mapping = NULL,
  data = NULL,
  step = 500,
  label = NULL,
  label_loc = NULL,
  label_fun = NULL,
  label_parse = FALSE,
  label_style = NULL,
  check_overlap = FALSE,
  ...,
  na.rm = FALSE,
  inherit.aes = TRUE
)

geom_grid_specvol(
  mapping = NULL,
  data = NULL,
  step = 0.01,
  label = NULL,
  label_loc = NULL,
  label_fun = NULL,
  label_parse = FALSE,
  label_style = NULL,
  check_overlap = FALSE,
  ...,
  na.rm = FALSE,
  inherit.aes = TRUE
)

geom_grid_enthalpy(
  mapping = NULL,
  data = NULL,
  step = 20,
  label = NULL,
  label_loc = NULL,
  label_fun = NULL,
  label_parse = FALSE,
  label_style = NULL,
  check_overlap = FALSE,
  ...,
  na.rm = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Draw constant psychrometric properties grids
}
\details{
\itemize{
\item \code{geom_line_sat} for saturation line.
\item \code{geom_grid_relhum} for relative humidity grid in range \link{0, 100} in \%
\item \code{geom_grid_wetbulb} for wet-bulb temperature grid in °F \link{IP} or °C \link{SI}
\item \code{geom_grid_vappres} for partial pressure grid of water vapor in Psi \link{IP}
or Pa \link{SI}
\item \code{geom_grid_specvol} for specific volume grid in ft3 lb-1 of dry air \link{IP} or
in m3 kg-1 of dry air \link{SI}
\item \code{geom_grid_enthalpy} for moist air enthalpy grid in Btu lb-1 \link{IP} or kJ kg-1
}

What these \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} objects do are to take input values,
calculate the corresponding humidity ratio and replace the \code{y} aesthetic
values in each group.
}
