% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-grid.R
\name{geom_line_sat}
\alias{geom_line_sat}
\title{Draw saturation line}
\usage{
geom_line_sat(
  mapping = NULL,
  data = NULL,
  units = waiver(),
  pres = waiver(),
  n = 201,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{units}{A single string indicating the units sytem to use. Should be
either \code{"SI"} or \verb{"IP" or }waiver()\verb{which uses the value from the parent plot. Default:}waiver()`}

\item{pres}{A single number indicating the atmosphere pressure in Pa [SI] or
Psi [IP]. If \code{waiver()}, the pressure calculated from the parent
plot's altitude value will be used. Default: \code{waiver()}}

\item{n}{Number of points to interpolate along}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
\code{geom_line_sat()} draws a saturation line based on current psychrometric
chart's dry-bulb temperature (x axis) range and humidity ratio (y axis)
range.
}
\details{
\code{geom_line_sat()} is based on \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}}, so you can further
customize the line style in the same way.

Normally there is no need to add another saturation line since \code{\link[=ggpsychro]{ggpsychro()}}
calls \code{geom_line_sat()} internally and makes sure that it is always rendered
at the last.
}
\section{Aesthetics}{


\code{geom_line_sat()} is drawing using \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} so support the
same aesthetics: \code{alpha}, \code{color}, \code{linetype} and \code{size}. It also has
aesthetics that control the calculation of the saturation line points
(required aesthetics are in bold):
}

\examples{
# by default, a saturation line is automatically added when calling 'ggpsychro()' function
ggpsychro()

# you can add another saturation line
ggpsychro() + geom_line_sat(units = "SI", pres = 101325, color = "blue", size = 2)

}
\seealso{
\code{\link[=ggpsychro]{ggpsychro()}}
}
