% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto-classes.R, R/stat.R
\docType{data}
\name{ggpsychro-extensions}
\alias{ggpsychro-extensions}
\alias{StatRelhum}
\alias{StatWetbulb}
\alias{StatVappres}
\alias{StatSpecvol}
\alias{StatEnthalpy}
\title{ggpsychro extensions to ggplot2}
\description{
ggpsychro makes heavy use of the ggproto class system to extend the
functionality of ggplot2. In general the actual classes should be of little
interest to users as the standard ggplot2 api of using geom_* and stat_*
functions for building up the plot is encouraged.
}
\keyword{datasets}
\keyword{internal}
