% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro-package.R
\docType{package}
\name{ggpsychro-package}
\alias{ggpsychro-package}
\title{ggpsychro: A 'ggplot2' Extension for Making Psychrometric Charts}
\description{
Provides 'ggplot2' coordinates, layers, scales, themes, and presets for creating psychrometric charts. The package supports SI and IP units, psychrometric grids, state points, process lines, zones, and thermal comfort overlays for heating, ventilation, air conditioning, and building performance workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hongyuanjia.github.io/ggpsychro/}
  \item \url{https://github.com/hongyuanjia/ggpsychro}
  \item Report bugs at \url{https://github.com/hongyuanjia/ggpsychro/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID}) [copyright holder]

}
