% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro-package.R
\docType{package}
\name{ggpsychro-package}
\alias{ggpsychro-package}
\alias{_PACKAGE}
\title{ggpsychro: A 'ggplot2' Extension for Making Pyschrometric Charts}
\description{
A 'ggplot2' extension for making pyschrometric charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hongyuanjia/ggpsychro}
  \item Report bugs at \url{https://github.com/hongyuanjia/ggpsychro/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuan.jia@bears-berkeley.sg} (\href{https://orcid.org/02-0075-8183}{ORCID})

}
