% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro.R
\name{ggpsychro}
\alias{ggpsychro}
\title{Create a ggpsychro plot}
\usage{
ggpsychro(
  data = NULL,
  mapping = aes(),
  tdb_lim = NULL,
  hum_lim = NULL,
  altitude = 0L,
  units = "SI",
  mollier = FALSE
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{ggplot2::fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, must be supplied in each layer added to the plot.}

\item{tdb_lim}{A numeric vector of length-2 indicating the dry-bulb
temperature limits. Should be in range
\verb{[-50, 100]} degree_C [SI] or
\verb{[-58, 212]} degree_F [IP].
If \code{NULL}, trained data ranges will be used when available, otherwise
a default display range will be used. Default: \code{NULL}.}

\item{hum_lim}{A numeric vector of length-2 indicating the humidity ratio
limits. Should be in range
\verb{[0, 60]} g_H20 kg_Air-1 [SI] or
\verb{[0, 420]} gr_H20 lb_Air-1 [IP].
If \code{NULL}, trained data ranges will be used when available, otherwise
a default display range will be used. Default: \code{NULL}.}

\item{altitude}{A single number of altitude in m [SI] or ft [IP]. Default:
\code{0}.}

\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}.}

\item{mollier}{If \code{TRUE}, a Mollier chart will be created instead of a
psychrometric chart. Default: \code{FALSE}.}
}
\value{
An object of class \code{gg} onto which layers, scales, etc. can be added.
}
\description{
\code{ggpsychro()} creates a ggplot object configured for psychrometric charts.
It stores chart metadata, installs \code{\link[=coord_psychro]{coord_psychro()}}, applies default axis
labels, and adds \code{\link[=theme_psychro]{theme_psychro()}} so psychrometric grids, states, zones, and
comfort overlays can be added with the usual ggplot2 \code{+} workflow.
}
\examples{
ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 50))

ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 50), mollier = TRUE)

ggpsychro(tdb_lim = c(0, 50), hum_lim = c(0, 50), units = "IP", altitude = 100)

}
\author{
Hongyuan Jia
}
\keyword{psychrometric}
