% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsychro.R
\name{ggpsychro}
\alias{ggpsychro}
\title{Create a ggpsychro plot}
\usage{
ggpsychro(
  data = NULL,
  mapping = aes(),
  tdb_lim = c(0, 50),
  hum_lim = c(0, 50),
  altitude = 0L,
  mask_style = waiver(),
  sat_style = waiver(),
  units = "SI",
  mollier = FALSE
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{ggplot2::fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, must be supplied in each layer added to the plot.}

\item{tdb_lim}{A numeric vector of length-2 indicating the dry-bulb
temperature limits. Should be in range
\verb{[-50, 100]} 째C [SI] or
\verb{[-58, 212]} 째F [IP].
If \code{waiver()}, default values will be
\verb{[0, 50]} 째C [SI] or
\verb{[30, 120]} 째F [IP]. Default: \code{waiver()}.}

\item{hum_lim}{A numeric vector of length-2 indicating the humidity ratio
limits. Should be in range
\verb{[0, 60]} g_H20 kg_Air-1 [SI] or
\verb{[0, 350]} gr_H20 lb_Air-1 [IP].
If \code{waiver()}, default values will be
\verb{[0, 30]} g_H20 kg_Air-1 [SI] or
\verb{[0, 210]} gr_H20 lb_Air-1 [IP]. Default: \code{waiver()}.}

\item{altitude}{A single number of altitude in m [SI] or ft [IP].}

\item{mask_style}{A list containg settings to format mask area. Will be
directly passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}. If \code{waiver()}, defaults
below will be used:
\itemize{
\item \code{fill}: \code{white}
\item \code{linetype}: \code{1}
\item \code{size}: \code{0.5}
}}

\item{sat_style}{A list containg settings to format saturation line. Will be
directly passed to \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}}. If \code{waiver()}, defaults
below will be used:
\itemize{
\item \code{color}: \verb{#DA251D}
\item \code{linetype}: \code{1}
\item \code{size}: \code{1}
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.}

\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}.}

\item{mollier}{If \code{TRUE}, a Mollier chart will be created instead of a
psychrometric chart. Default: \code{FALSE}.}
}
\value{
An object of class \code{gg} onto which layers, scales, etc. can be added.
}
\description{
This function is the equivalent of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} in ggplot2.
It takes care of setting up the plot object along with creating the layout
for the plot based on the graph and the specification passed in.
Alternatively a layout can be prepared in advance using
\code{create_layout} and passed as the data argument. See \emph{Details} for
a description of all available layouts.
}
\examples{
ggpsychro()

}
\author{
Hongyuan Jia
}
\keyword{psychrometric}
