% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_drybulb_continuous}
\alias{scale_drybulb_continuous}
\alias{scale_humratio_continuous}
\alias{scale_grid_relhum}
\alias{scale_grid_wetbulb}
\alias{scale_grid_vappres}
\alias{scale_grid_specvol}
\alias{scale_grid_enthalpy}
\title{Transformation object for psychrometric chart}
\usage{
scale_drybulb_continuous(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  units = "SI",
  ...
)

scale_humratio_continuous(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  units = "SI",
  ...
)

scale_grid_relhum(
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  units = "SI",
  position = "right",
  ...
)

scale_grid_wetbulb(
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  units = "SI",
  position = "right",
  ...
)

scale_grid_vappres(
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  units = "SI",
  position = "right",
  ...
)

scale_grid_specvol(
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  units = "SI",
  position = "right",
  ...
)

scale_grid_enthalpy(
  breaks = waiver(),
  minor_breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  units = "SI",
  position = "right",
  ...
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
transformation object
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output
}}

\item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (one minor break between
each major break)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks.
}}

\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits
}}

\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}.}

\item{...}{Other arguments passed on to \code{scale_(x|y)_continuous()}}

\item{position}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}
}
\description{
Transformation object for psychrometric chart
}
