% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_relhum}
\alias{stat_relhum}
\alias{stat_wetbulb}
\alias{stat_vappres}
\alias{stat_specvol}
\alias{stat_enthalpy}
\title{Draw constant-property psychrometric data}
\usage{
stat_relhum(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_wetbulb(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_vappres(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_specvol(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_enthalpy(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}
}
\value{
A ggplot layer.
}
\description{
Draw constant-property psychrometric data
}
\details{
These stats convert common psychrometric properties to the humidity-ratio
coordinate used by the chart. Use them when the data records dry-bulb
temperature plus another property instead of humidity ratio.

Required input aesthetics:
\itemize{
\item \code{stat_relhum()} uses \code{x} and \code{relhum}; \code{relhum} is relative humidity in
percent from 0 to 100.
\item \code{stat_wetbulb()} uses \code{x} and \code{wetbulb}; wet-bulb temperature follows the
parent chart units.
\item \code{stat_vappres()} uses \code{x} and \code{vappres}; vapor pressure is Pa in SI and Psi
in IP.
\item \code{stat_specvol()} uses \code{x} and \code{specvol}; specific volume is m^3/kg dry air
in SI and ft^3/lb dry air in IP.
\item \code{stat_enthalpy()} uses \code{x} and \code{enthalpy}; enthalpy is J/kg dry air in SI
and Btu/lb dry air in IP.
}

When used with \code{\link[=ggpsychro]{ggpsychro()}}, \code{units} and pressure are inherited from the
plot. If these stats are used directly inside ordinary ggplot2 geoms via
\code{stat = "relhum"} or similar, supply both \code{units} and \code{pres} explicitly.

Internally, each stat computes humidity ratio and replaces the layer \code{y}
aesthetic before the psychrometric coordinate transform is applied.
}
\examples{
states <- data.frame(
    tdb = c(18, 22, 26, 30),
    relhum = c(70, 55, 45, 35)
)

ggpsychro(tdb_lim = c(10, 35), hum_lim = c(0, 25)) +
    stat_relhum(aes(x = tdb, relhum = relhum), data = states)

wetbulb_line <- data.frame(tdb = 18:30, wetbulb = 16)
ggpsychro(tdb_lim = c(10, 35), hum_lim = c(0, 25)) +
    geom_psychro_grid_wetbulb() +
    stat_wetbulb(aes(x = tdb, wetbulb = wetbulb),
        data = wetbulb_line, geom = "line")

# The stats can also be used from ordinary ggplot2 geoms.
ggpsychro(tdb_lim = c(10, 35), hum_lim = c(0, 25)) +
    geom_psychro_grid_wetbulb() +
    geom_line(aes(x = tdb, wetbulb = wetbulb),
        data = wetbulb_line, stat = "wetbulb")

vapour_pressure <- data.frame(
    tdb = c(12, 18, 24, 30),
    vappres = c(900, 1200, 1800, 2400)
)
ggpsychro(tdb_lim = c(10, 35), hum_lim = c(0, 25)) +
    stat_vappres(aes(x = tdb, vappres = vappres),
        data = vapour_pressure)

specific_volume <- data.frame(
    tdb = c(20, 25, 30),
    specvol = c(0.84, 0.86, 0.88)
)
ggpsychro(tdb_lim = c(10, 35), hum_lim = c(0, 25)) +
    stat_specvol(aes(x = tdb, specvol = specvol),
        data = specific_volume)

enthalpy <- data.frame(
    tdb = c(18, 24, 30),
    enthalpy = c(35000, 50000, 65000)
)
ggpsychro(tdb_lim = c(10, 35), hum_lim = c(0, 25)) +
    stat_enthalpy(aes(x = tdb, enthalpy = enthalpy), data = enthalpy)

}
