% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_relhum}
\alias{stat_relhum}
\alias{stat_wetbulb}
\alias{stat_vappres}
\alias{stat_specvol}
\alias{stat_enthalpy}
\title{Calculate psychrometric properties of moist air}
\usage{
stat_relhum(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_wetbulb(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_vappres(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_specvol(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_enthalpy(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Calculate psychrometric properties of moist air
}
\details{
\itemize{
\item \code{stat_relhum} requires an extra \code{relhum} aesthetics for relative humidity
in range \link{0, 100} in \%
\item \code{stat_wetbulb} requires an extra \code{wetbulb} aesthetics for wet-bulb
temperature in °F \link{IP} or °C \link{SI}
\item \code{stat_vappres} requires an extra \code{vappres} aesthetics for partial pressure
of water vapor in moist air in Psi \link{IP} or Pa \link{SI}
\item \code{stat_specvol} requires an extra \code{specvol} aesthetics for specific volume
of moist air in ft3 lb-1 of dry air \link{IP} or in m3 kg-1 of dry air \link{SI}
\item \code{stat_enthalpy} requires an extra \code{enthalpy} aesthetics for moist air
enthalpy in Btu lb-1 \link{IP} or J kg-1
}

What these \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} objects do are to take input values,
calculate the corresponding humidity ratio and replace the \code{y} aesthetic
values in each group.

All of stats above requires two additional aesthetics:
\itemize{
\item \code{units}: A single string indicating the units sytem to use. Should be
either \code{"SI"} or \verb{"IP" or }waiver()\verb{which uses the value from the parent plot. Default:}waiver()`
\item \code{pres}: A single number indicating the atmosphere pressure in Pa \link{SI} or
Psi \link{IP}. If \code{waiver()}, the pressure calculated from the parent plot's
altitude value will be used. Default: \code{waiver()}
}

However, when these stats are used inside a ggplot \verb{geom_*} as the \code{stat}
argument, both \code{units} and \code{pres} have to be specified.
}
\examples{
p <- ggpsychro() + geom_grid_relhum() # add relative humidity grid lines

# draw a point with dry-bulb at 30C and relative humidity at 60\%
p + geom_point(aes(x = 30, relhum = 0.6), stat = "relhum", size = 5)

# draw a constant relative humidity line of 60\% with dry-bulb from 20C to 30C
## use stat_* directly
p + stat_relhum(geom = "line", aes(x = 20:30, relhum = 0.6), size = 2)
## OR
## use as thet `stat` argument inside an ggplot `geom_*` function
p + geom_line(aes(x = 20:30, relhum = 0.6), stat = "relhum")

}
