% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\alias{theme_grey_psychro}
\alias{theme_gray_psychro}
\alias{theme_psychro}
\title{Custom theme for psychrometric chart.}
\usage{
theme_grey_psychro(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_gray_psychro(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_psychro(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
A ggplot2 theme.
}
\description{
Custom theme for psychrometric chart.
}
\examples{
theme_psychro()

}
\author{
Hongyuan Jia
}
