% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_psychro}
\alias{theme_psychro}
\title{Custom theme for psychrometric chart.}
\usage{
theme_psychro(
  background = "white",
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{asp}{Aspect ratio of plot. Defaults to NULL.}
}
\value{
A ggplot2 theme.
}
\description{
Custom theme for psychrometric chart.
}
\examples{
theme_psychro()

}
\author{
Hongyuan Jia
}
\keyword{internal}
