% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{drybulb_trans}
\alias{drybulb_trans}
\alias{humratio_trans}
\alias{relhum_trans}
\alias{wetbulb_trans}
\alias{vappres_trans}
\alias{specvol_trans}
\alias{enthalpy_trans}
\title{Create transformation objects for psychrometric chart}
\usage{
drybulb_trans(units)

humratio_trans(units)

relhum_trans(units)

wetbulb_trans(units)

vappres_trans(units)

specvol_trans(units)

enthalpy_trans(units)
}
\arguments{
\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}.}
}
\description{
Create transformation objects for psychrometric chart
}
